Refactor parquet::arrow module
